Update hilo.py
I have made a bunch of changes. I made the code clean so that it could be easily read. The only part that is not working is the "input" in "get_inputs(self)". Somehow it is not being stored as a string value or not stored at all. I tried debugging it but it didn't show anything.